% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{animate_css}
\alias{animate_css}
\alias{use_animate_css}
\alias{html_dependency_animate_css}
\title{Animate.css}
\usage{
use_animate_css(minified = FALSE, xaringan = TRUE)

html_dependency_animate_css(minified = FALSE, xaringan = TRUE)
}
\arguments{
\item{minified}{Should the minified or full CSS source be used?}

\item{xaringan}{When \code{TRUE}, the \code{.animated} selector is modified so that the
animation is only applied when the slide is visible. Without this,
presentations with many animated slides may have poor performance,
especially on page load. Set to \code{FALSE} to use animate.css in other
HTML-based documents.}
}
\value{
An \link[htmltools:tagList]{htmltools::tagList} with the tile view dependencies, or an
\link[htmltools:htmlDependency]{htmltools::htmlDependency}.
}
\description{
Animate.css is a popular collection of CSS animations. It contains "a
bunch of cool, fun, and cross-browser animations for you to use in your
projects. Great for emphasis, home pages, sliders, and general
just-add-water-awesomeness."
}
\section{Functions}{
\itemize{
\item \code{use_animate_css}: Use animate.css in your xaringan slides.

\item \code{html_dependency_animate_css}: Returns an \link[htmltools:htmlDependency]{htmltools::htmlDependency} with the tile
view dependencies. Most users will want to use \code{use_animate_css()}.
}}

\section{Usage}{
 To add animate.css to your xaringan presentation, add the
following code chunk to your slides' R Markdown file.\preformatted{```{r xaringan-animate, echo=FALSE}
xaringanExtra::use_animate_css()
```

---
class: animated fadeInLeft

This slide fades in from the left!
}
}

\references{
See \href{http://daneden.github.io/animate.css}{animate.css} for a
full list of animations.
}
