% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canvas.R
\name{canvas}
\alias{canvas}
\alias{use_canvas}
\alias{html_dependency_canvas}
\alias{html_dependency_fontawesome}
\title{Canvas}
\usage{
use_canvas(
  default_color = "#000000",
  pen_size = 2,
  eraser_size = pen_size * 10
)

html_dependency_canvas()

html_dependency_fontawesome()
}
\arguments{
\item{default_color}{Pen color (default is "#000000"). Users may choose any
valid CSS hex color.}

\item{pen_size}{Pen size (default is 2).}

\item{eraser_size}{Eraser size (default is 10 times \code{pen_size}).}
}
\value{
An \code{htmltools::tagList()} with the fabric dependencies, or an
\code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}.
}
\description{
Canvas provides a simple set of tools to allow you to mark up
your \code{xaringan} slides. A toolbox will appear in the top-right corner once
you begin navigating your slides. Click the edit icon to begin. Use the
eraser to remove select markup, or the trash can to clear the entire
canvas. When you are done, click the green "✓" to keep the markup, or the
red "X" to hide the canvas.
}
\details{
You may toggle the visibility of the canvas toolbox by pressing
\strong{D} at any time. Markup will persist when changing slides as long as the
canvas isn't cleared. You may save a permanent copy of the slides with the
markup by saving your presentation (e.g. using Chrome > File > Print).
}
\section{Functions}{
\itemize{
\item \code{use_canvas}: Adds a markup canvas to your xaringan slides.

\item \code{html_dependency_canvas}: Returns an \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}} with the
canvas dependencies for use in xaringan and R Markdown documents. Most
users will want to use \code{use_canvas()} instead.

\item \code{html_dependency_fontawesome}: Returns an \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}} with the
fontawesome dependencies for use in xaringan and R Markdown documents.
Most users will want to use \code{use_canvas()} instead.
}}

\section{Usage}{
 To add a drawing canvas to your xaringan presentation,
add the following code chunk to your slides' R Markdown file.\if{html}{\out{<div class="markdown">}}\preformatted{```\{r xaringan-canvas, echo=FALSE\}
xaringanExtra::use_canvas()
```
}\if{html}{\out{</div>}}
}

