% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excitation.R
\name{excitation}
\alias{excitation}
\alias{use_excitation}
\alias{excite}
\alias{html_dependency_excitation}
\alias{html_dependency_popper}
\alias{html_dependency_tippy}
\title{Excitation}
\usage{
use_excitation()

excite(key)

html_dependency_excitation()

html_dependency_popper()

html_dependency_tippy()
}
\description{
Excitation lets you add visual citations as tooltips to your
\pkg{xaringan} slides. Simply include a \code{.bib} file in your R Markdown
YAML header, as usual, and cite away!
}
\section{Functions}{
\itemize{
\item \code{use_excitation}: Adds visual citations to your xaringan slides.

\item \code{excite}: Adds visual citations to your xaringan slides.

\item \code{html_dependency_excitation}: Returns an \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}} with the
\code{excitation.js} dependencies for use in xaringan and R Markdown
documents. Most users will want to use \code{use_excitation()} instead.

\item \code{html_dependency_popper}: Returns an \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}} with the
\code{popper.js} dependencies for use in xaringan and R Markdown
documents. Most users will want to use \code{use_excitation()} instead.

\item \code{html_dependency_tippy}: Returns an \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}} with the
\code{tippy.js} dependencies for use in xaringan and R Markdown
documents. Most users will want to use \code{use_excitation()} instead.
}}

\section{Usage}{
 To use excitation for visual citations, add the following
code chunk to your slides' R Markdown file.

\if{html}{\out{<div class="sourceCode markdown">}}\preformatted{```\{r xaringan-excitation, echo=FALSE\}
xaringanExtra::use_excitation()
```
}\if{html}{\out{</div>}}
}

\examples{
use_excitation()

}
