% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabric.R
\name{fabric}
\alias{fabric}
\alias{use_fabric}
\alias{html_dependency_fabric}
\title{Fabric Canvas}
\usage{
use_fabric()

html_dependency_fabric()
}
\value{
An \code{htmltools::tagList()} with the fabric dependencies, or an
\code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}.
}
\description{
Fabric gives you a way to add a canvas on top of your slide to draw on.
Just press \strong{D} at any point in your slideshow and a transparent
drawing canvas will be added to your slide. Drawing tools will appear in
the top right corner of the slide to control stroke color and thickness.
}
\details{
Drawings will persist when changing slides as long as the canvas
isn't removed. The canvas can be removed at any time by pressing \strong{D}.
You may also keep a permanent copy of the slides with the drawings by
saving your presentation (e.g. using Chrome > File > Print).
}
\section{Functions}{
\itemize{
\item \code{use_fabric}: Adds a drawing canvas to your xaringan slides.

\item \code{html_dependency_fabric}: Returns an \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}} with the fabric
dependencies. Most users will want to use \code{use_fabric()}.
}}

\section{Usage}{
 To add a drawing canvas to your xaringan presentation,
add the following code chunk to your slides' R Markdown file.\if{html}{\out{<div class="markdown">}}\preformatted{```\{r xaringan-fabric, echo=FALSE\}
xaringanExtra::use_fabric()
```
}\if{html}{\out{</div>}}
}

