% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashcard.R
\name{flashcard}
\alias{flashcard}
\title{Create a Flashcard Component}
\usage{
flashcard(front, back)
}
\arguments{
\item{front}{the HTML conent to put on the front of the card}

\item{back}{the HTML conent to put on the back of the card}
}
\value{
a div that on hover you can switch between front and back
}
\description{
Create a Flashcard Component
}
