% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scribble.R
\name{scribble}
\alias{scribble}
\alias{use_scribble}
\alias{html_dependency_fabricjs}
\alias{html_dependency_scribble}
\title{Scribble}
\usage{
use_scribble(pen_color = "#FF0000", pen_size = 3, eraser_size = pen_size * 10)

html_dependency_fabricjs(minimized = TRUE)

html_dependency_scribble(pen_color, pen_size, eraser_size)
}
\arguments{
\item{pen_color}{Initial pen color (default is \verb{"#FF0000} (red)). May be any
valid CSS Hex, RGB, or HSL color.}

\item{pen_size}{Pen size (default is 3).}

\item{eraser_size}{Eraser size (default is \code{pen_size * 10}).}

\item{minimized}{Use the minimized \code{fabric.js} dependency?}
}
\value{
An \code{htmltools::tagList()} with the scribble dependencies, or an
\code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}.
}
\description{
\code{Scribble} provides a set of tools to allow you to mark up
your \code{xaringan} slides. Click the "edit" icon to begin drawing. Use the
eraser to remove select markup, or the trash to clear the entire
canvas. NOTE: You will not be able to navigate slides while in draw or
erase mode.
}
\details{
You may toggle the visibility of the \code{Scribble} toolbox by pressing
\strong{S} at any time. Markup will persist when changing slides as long as the
canvas isn't cleared. You may save a permanent copy of the slides with the
markup by saving your presentation (e.g. using Chrome > File > Print).
}
\section{Functions}{
\itemize{
\item \code{use_scribble}: Adds \code{Scribble} to your xaringan slides.

\item \code{html_dependency_fabricjs}: Returns an \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}} with the
\code{fabric.js} dependencies for use in xaringan and R Markdown documents.
Most users will want to use \code{use_scribble()} instead.

\item \code{html_dependency_scribble}: Returns an \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}} with the
\code{scribble} dependencies for use in xaringan and R Markdown documents. Most
users will want to use \code{use_scribble()} instead.
}}

\section{Usage}{
 To add \code{Scribble} to your \code{xaringan} presentation,
add the following code chunk to your slides' R Markdown file.\if{html}{\out{<div class="markdown">}}\preformatted{```\{r xaringan-scribble, echo=FALSE\}
xaringanExtra::use_scribble()
```
}\if{html}{\out{</div>}}
}

