% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide-tone.R
\name{slide_tone}
\alias{slide_tone}
\alias{xaringan_slide_tone}
\alias{xaringan_slide_tone_dependency}
\title{Slide Tone}
\usage{
xaringan_slide_tone()

xaringan_slide_tone_dependency()
}
\value{
An \link[htmltools:tagList]{htmltools::tagList} with the slide tone dependencies, or an
\link[htmltools:htmlDependency]{htmltools::htmlDependency}.
}
\description{
Slide tone plays a subtle sound when you change slides.The tones increase in
pitch for each slide from a low C to a high C note. The tone pitch stays the
same for incremental slides.
}
\section{Functions}{
\itemize{
\item \code{xaringan_slide_tone}: Adds slide tone to your xaringan slides.

\item \code{xaringan_slide_tone_dependency}: Returns an \link[htmltools:htmlDependency]{htmltools::htmlDependency} with the tile
view dependencies. Most users will want to use \code{xaringan_slide_tone()}.
}}

\section{Usage}{
 To add slide tone to your xaringan presentation, add the
following code chunk to your slides' R Markdown file.\preformatted{```{r xaringan-slide-tone, echo=FALSE}
xaringanExtra::xaringan_slide_tone()
```
}
}

\references{
\href{https://tonejs.github.io/}{tone.js}
}
