% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tachyons.R
\name{style_tachyons}
\alias{style_tachyons}
\title{Style Tachyons}
\usage{
style_tachyons(
  spacing_none = NULL,
  spacing_extra_small = NULL,
  spacing_small = NULL,
  spacing_medium = NULL,
  spacing_large = NULL,
  spacing_extra_large = NULL,
  spacing_extra_extra_large = NULL,
  spacing_extra_extra_extra_large = NULL,
  black = NULL,
  near_black = NULL,
  dark_gray = NULL,
  mid_gray = NULL,
  gray = NULL,
  silver = NULL,
  light_silver = NULL,
  moon_gray = NULL,
  light_gray = NULL,
  near_white = NULL,
  white = NULL,
  transparent = NULL,
  black_90 = NULL,
  black_80 = NULL,
  black_70 = NULL,
  black_60 = NULL,
  black_50 = NULL,
  black_40 = NULL,
  black_30 = NULL,
  black_20 = NULL,
  black_10 = NULL,
  black_05 = NULL,
  black_025 = NULL,
  black_0125 = NULL,
  white_90 = NULL,
  white_80 = NULL,
  white_70 = NULL,
  white_60 = NULL,
  white_50 = NULL,
  white_40 = NULL,
  white_30 = NULL,
  white_20 = NULL,
  white_10 = NULL,
  white_05 = NULL,
  white_025 = NULL,
  white_0125 = NULL,
  dark_red = NULL,
  red = NULL,
  light_red = NULL,
  orange = NULL,
  gold = NULL,
  yellow = NULL,
  light_yellow = NULL,
  purple = NULL,
  light_purple = NULL,
  dark_pink = NULL,
  hot_pink = NULL,
  pink = NULL,
  light_pink = NULL,
  dark_green = NULL,
  green = NULL,
  light_green = NULL,
  navy = NULL,
  dark_blue = NULL,
  blue = NULL,
  light_blue = NULL,
  lightest_blue = NULL,
  washed_blue = NULL,
  washed_green = NULL,
  washed_yellow = NULL,
  washed_red = NULL
)
}
\description{
Creates a \verb{<style>} tag to set CSS variables used by Tachyons.
}
\seealso{
tachyons
}
\arguments{ \item{spacing_none}{\code{--spacing-none}, default is \code{0}}
\item{spacing_extra_small}{\code{--spacing-extra-small}, default is \code{.25rem}}
\item{spacing_small}{\code{--spacing-small}, default is \code{.5rem}}
\item{spacing_medium}{\code{--spacing-medium}, default is \code{1rem}}
\item{spacing_large}{\code{--spacing-large}, default is \code{2rem}}
\item{spacing_extra_large}{\code{--spacing-extra-large}, default is \code{4rem}}
\item{spacing_extra_extra_large}{\code{--spacing-extra-extra-large}, default is \code{8rem}}
\item{spacing_extra_extra_extra_large}{\code{--spacing-extra-extra-extra-large}, default is \code{16rem}}
\item{black}{\code{--black}, default is \code{#000}}
\item{near_black}{\code{--near-black}, default is \code{#111}}
\item{dark_gray}{\code{--dark-gray}, default is \code{#333}}
\item{mid_gray}{\code{--mid-gray}, default is \code{#555}}
\item{gray}{\code{--gray}, default is \code{#777}}
\item{silver}{\code{--silver}, default is \code{#999}}
\item{light_silver}{\code{--light-silver}, default is \code{#aaa}}
\item{moon_gray}{\code{--moon-gray}, default is \code{#ccc}}
\item{light_gray}{\code{--light-gray}, default is \code{#eee}}
\item{near_white}{\code{--near-white}, default is \code{#f4f4f4}}
\item{white}{\code{--white}, default is \code{#fff}}
\item{transparent}{\code{--transparent}, default is \code{transparent}}
\item{black_90}{\code{--black-90}, default is \code{rgba( 0, 0, 0, .9 )}}
\item{black_80}{\code{--black-80}, default is \code{rgba( 0, 0, 0, .8 )}}
\item{black_70}{\code{--black-70}, default is \code{rgba( 0, 0, 0, .7 )}}
\item{black_60}{\code{--black-60}, default is \code{rgba( 0, 0, 0, .6 )}}
\item{black_50}{\code{--black-50}, default is \code{rgba( 0, 0, 0, .5 )}}
\item{black_40}{\code{--black-40}, default is \code{rgba( 0, 0, 0, .4 )}}
\item{black_30}{\code{--black-30}, default is \code{rgba( 0, 0, 0, .3 )}}
\item{black_20}{\code{--black-20}, default is \code{rgba( 0, 0, 0, .2 )}}
\item{black_10}{\code{--black-10}, default is \code{rgba( 0, 0, 0, .1 )}}
\item{black_05}{\code{--black-05}, default is \code{rgba( 0, 0, 0, .05 )}}
\item{black_025}{\code{--black-025}, default is \code{rgba( 0, 0, 0, .025 )}}
\item{black_0125}{\code{--black-0125}, default is \code{rgba( 0, 0, 0, .0125 )}}
\item{white_90}{\code{--white-90}, default is \code{rgba( 255, 255, 255, .9 )}}
\item{white_80}{\code{--white-80}, default is \code{rgba( 255, 255, 255, .8 )}}
\item{white_70}{\code{--white-70}, default is \code{rgba( 255, 255, 255, .7 )}}
\item{white_60}{\code{--white-60}, default is \code{rgba( 255, 255, 255, .6 )}}
\item{white_50}{\code{--white-50}, default is \code{rgba( 255, 255, 255, .5 )}}
\item{white_40}{\code{--white-40}, default is \code{rgba( 255, 255, 255, .4 )}}
\item{white_30}{\code{--white-30}, default is \code{rgba( 255, 255, 255, .3 )}}
\item{white_20}{\code{--white-20}, default is \code{rgba( 255, 255, 255, .2 )}}
\item{white_10}{\code{--white-10}, default is \code{rgba( 255, 255, 255, .1 )}}
\item{white_05}{\code{--white-05}, default is \code{rgba( 255, 255, 255, .05 )}}
\item{white_025}{\code{--white-025}, default is \code{rgba( 255, 255, 255, .025 )}}
\item{white_0125}{\code{--white-0125}, default is \code{rgba( 255, 255, 255, .0125 )}}
\item{dark_red}{\code{--dark-red}, default is \code{#e7040f}}
\item{red}{\code{--red}, default is \code{#ff4136}}
\item{light_red}{\code{--light-red}, default is \code{#ff725c}}
\item{orange}{\code{--orange}, default is \code{#ff6300}}
\item{gold}{\code{--gold}, default is \code{#ffb700}}
\item{yellow}{\code{--yellow}, default is \code{#ffd700}}
\item{light_yellow}{\code{--light-yellow}, default is \code{#fbf1a9}}
\item{purple}{\code{--purple}, default is \code{#5e2ca5}}
\item{light_purple}{\code{--light-purple}, default is \code{#a463f2}}
\item{dark_pink}{\code{--dark-pink}, default is \code{#d5008f}}
\item{hot_pink}{\code{--hot-pink}, default is \code{#ff41b4}}
\item{pink}{\code{--pink}, default is \code{#ff80cc}}
\item{light_pink}{\code{--light-pink}, default is \code{#ffa3d7}}
\item{dark_green}{\code{--dark-green}, default is \code{#137752}}
\item{green}{\code{--green}, default is \code{#19a974}}
\item{light_green}{\code{--light-green}, default is \code{#9eebcf}}
\item{navy}{\code{--navy}, default is \code{#001b44}}
\item{dark_blue}{\code{--dark-blue}, default is \code{#00449e}}
\item{blue}{\code{--blue}, default is \code{#357edd}}
\item{light_blue}{\code{--light-blue}, default is \code{#96ccff}}
\item{lightest_blue}{\code{--lightest-blue}, default is \code{#cdecff}}
\item{washed_blue}{\code{--washed-blue}, default is \code{#f6fffe}}
\item{washed_green}{\code{--washed-green}, default is \code{#e8fdf5}}
\item{washed_yellow}{\code{--washed-yellow}, default is \code{#fffceb}}
\item{washed_red}{\code{--washed-red}, default is \code{#ffdfdf}} }
