% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text-poster.R
\name{text_poster}
\alias{text_poster}
\alias{use_text_poster}
\alias{html_dependency_text_poster}
\title{Text Poster (Deprecated)}
\usage{
text_poster(
  text,
  width = "100\%",
  height = "100\%",
  padding = "1em",
  class = NULL,
  ...
)

use_text_poster()

html_dependency_text_poster()
}
\arguments{
\item{text}{The text to include in the poster text block. Include line breaks
to force new lines in the block.}

\item{width, height, padding}{The width, height, and padding of the poster text
container \verb{<div>}. Set to \code{NULL} to leave unset, but be sure to use global
CSS rules to set the width and height of the \code{.text-poster} container.}

\item{class}{Additional classes added to the poster text container \verb{<div>}.}

\item{...}{Additional named arguments added to the poster text container
\verb{<div>} as inline styles, e.g. \code{margin = "0 auto"}.}
}
\value{
The \code{text_poster()} function returns HTML with the following
structure:\if{html}{\out{<div class="sourceCode html">}}\preformatted{<div class="text-poster" style="width:100\%;height:100\%;padding:1em;">
  <div class="text-poster__text" data-text="first line\\\\nsecond line"></div>
</div>
}\if{html}{\out{</div>}}

If specified, additional classes are added to the \code{.text-poster} \code{div}.
Additional inline styles can be applied using named arguments to
\code{text_poster()}, and the pre-set styles can be disabled by setting them
equal to \code{NULL}. Note that the \code{div.text-poster} must have a width and
height.

Once rendered, the final output will have the following structure:\if{html}{\out{<div class="sourceCode html">}}\preformatted{<div class="text-poster" style="width:100\%;height:100\%;padding:1em;">
  <div class="text-poster__text" data-text="first line\\\\nsecond line">
    <div class="line-container">
      <div class="line">first line</div>
      <div class="line">second line</div>
    </div>
  </div>
</div>
}\if{html}{\out{</div>}}
}
\description{
DEPRECATED: Text poster has been deprecated and will be removed in a future
release. Text poster typesets text to fit within a rectangular bounding box,
with the text on each line scaled to fit the horizontal space.
}
\section{Functions}{
\itemize{
\item \code{text_poster}: Create poster text in your xaringan slides.

\item \code{use_text_poster}: Prepares your xaringan slides to use \code{text-poster}.
If you use \code{text_poster()} to create your poster-block text, you do not
need to call this function.

\item \code{html_dependency_text_poster}: Returns an \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}} with the
\code{text-poster} dependencies. Most users will want to use \code{use_text_poster()}.
}}

\section{Usage}{
 To add text-poster to your xaringan presentation, use
\code{text_poster()} to create a block of poster text. Include line breaks in
the text string to split the text block.\if{html}{\out{<div class="sourceCode markdown">}}\preformatted{```\{r poster-text, echo=FALSE\}`r ''`
xaringanExtra::text_poster(
  "There are no
  routine statistical
  questions, only questionable
  statistical routines."
)
```
}\if{html}{\out{</div>}}
}

\references{
\href{https://github.com/IMAGINARY/text-poster#readme}{text-poster}
}
