% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile-view.R
\name{tile_view}
\alias{tile_view}
\alias{xaringan_tile_view}
\alias{xaringan_tile_view_dependency}
\title{Tile View}
\usage{
xaringan_tile_view()

xaringan_tile_view_dependency()
}
\value{
An \link[htmltools:tagList]{htmltools::tagList} with the tile view dependencies, or an
\link[htmltools:htmlDependency]{htmltools::htmlDependency}.
}
\description{
Tile view gives you a way to quickly jump between slides. Just press \code{T} at
any point in your slideshow and the tile view appears. Click on a slide to
jump to the slide, or press \code{T} to exit tile view.
}
\section{Functions}{
\itemize{
\item \code{xaringan_tile_view}: Adds tile view to your xaringan slides.

\item \code{xaringan_tile_view_dependency}: Returns an \link[htmltools:htmlDependency]{htmltools::htmlDependency} with the tile
view dependencies. Most users will want to use \code{xaringan_tile_view()}.
}}

\section{Usage}{
 To add tile view to your xaringan presentation, add the
following code chunk to your slides' R Markdown file.\preformatted{```{r xaringan-tile-view, echo=FALSE}
xaringanExtra::xaringan_tile_view()
```
}
}

