% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hacker-type.R
\name{use_hacker_typing}
\alias{use_hacker_typing}
\alias{html_dependency_hacker_typing}
\title{Hacker Typing}
\usage{
use_hacker_typing()

html_dependency_hacker_typing()
}
\description{
FIXME
}
\section{Functions}{
\itemize{
\item \code{html_dependency_hacker_typing}: Returns an \link[htmltools:htmlDependency]{htmltools::htmlDependency} with the
hacker tpying dependencies. Most users will want to use
\code{use_hacker_typing()}.
}}

