% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress-bar.R
\name{use_progress_bar}
\alias{use_progress_bar}
\title{Add an animated progress bar}
\usage{
use_progress_bar(
  color = "red",
  location = c("top", "bottom"),
  height = "0.25em"
)
}
\arguments{
\item{color}{A valid CSS color to be used as the color of the progress bar.}

\item{location}{One of \code{"top"} or \code{"bottom"}.}

\item{height}{A valid CSS unit specifying the height of the progress bar.}
}
\value{
An \code{htmltools::tagList()} with the HTML dependencies required for
\strong{progress bar}.
}
\description{
Adds an animated progress bar to all slides.
}
\examples{
xaringanExtra::use_progress_bar("red", "top", "0.25em")

}
