% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabric.R
\name{fabric}
\alias{fabric}
\alias{use_fabric}
\alias{html_dependency_fabricjs}
\alias{html_dependency_fabric}
\title{Fabric}
\usage{
use_fabric(
  pen_color = "rgb(0, 0, 0)",
  pen_size = 3,
  eraser_color = "rgba(0, 0, 0, 0.1)",
  eraser_size = pen_size * 10
)

html_dependency_fabricjs()

html_dependency_fabric()
}
\arguments{
\item{pen_color}{Pen color (default is "black"). Users may choose any
valid CSS Hex, RGB, or HSL color.}

\item{pen_size}{Pen size (default is 3).}

\item{eraser_color}{Eraser color (default is "rgba(0, 0, 0, 0.1)").
Users may choose any valid CSS Hex, RGB, or HSL color.}

\item{eraser_size}{Eraser size (default is 10 times \code{pen_size}).}
}
\value{
An \code{htmltools::tagList()} with the fabric dependencies, or an
\code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}.
}
\description{
Fabric provides a set of tools to allow you to mark up
your \code{xaringan} slides. Click the "edit" icon to begin. Use the
eraser to remove select markup, or the trash can to clear the entire
canvas. NOTE: You will not be able to navigate slides while in draw or
erase mode.
}
\details{
You may toggle the visibility of the canvas toolbox by pressing
\strong{D} at any time. Markup will persist when changing slides as long as the
canvas isn't cleared. You may save a permanent copy of the slides with the
markup by saving your presentation (e.g. using Chrome > File > Print).
}
\section{Functions}{
\itemize{
\item \code{use_fabric}: Adds a markup canvas to your xaringan slides.

\item \code{html_dependency_fabricjs}: Returns an \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}} with the
fabric dependencies for use in xaringan and R Markdown documents. Most
users will want to use \code{use_fabric()} instead.

\item \code{html_dependency_fabric}: Returns an \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}} with the
fabric dependencies for use in xaringan and R Markdown documents. Most
users will want to use \code{use_fabric()} instead.
}}

\section{Usage}{
 To add a fabric canvases to your \code{xaringan} presentation,
add the following code chunk to your slides' R Markdown file.\if{html}{\out{<div class="markdown">}}\preformatted{```\{r xaringan-fabric, echo=FALSE\}
xaringanExtra::use_fabric()
```
}\if{html}{\out{</div>}}
}

