% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashcard.R
\name{flashcard}
\alias{flashcard}
\title{Create a Flashcard Component}
\usage{
flashcard(front, back)
}
\arguments{
\item{front, back}{The HTML content to put on the front or back of the card}
}
\value{
A \verb{<div>} that reveals content on \code{back} when hovering over \code{front}.
}
\description{
Create a Flashcard Component
}
