% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelset.R
\name{panelset}
\alias{panelset}
\alias{use_panelset}
\alias{html_dependency_panelset}
\title{Panelset}
\usage{
use_panelset()

html_dependency_panelset()
}
\value{
An \link[htmltools:tagList]{htmltools::tagList} with the panelset dependencies, or an
\link[htmltools:htmlDependency]{htmltools::htmlDependency}.
}
\description{
A panelset designed for showing off code, but useful for anything really.
}
\section{Functions}{
\itemize{
\item \code{use_panelset}: Adds panelset to your xaringan slides.

\item \code{html_dependency_panelset}: Returns an \link[htmltools:htmlDependency]{htmltools::htmlDependency} with the tile
view dependencies. Most users will want to use \code{use_panelset()}.
}}

\section{Usage}{
 To add panelset to your xaringan presentation, add the
following code chunk to your slides' R Markdown file.\if{html}{\out{<div class="markdown">}}\preformatted{```\{r xaringan-panelset, echo=FALSE\}
xaringanExtra::use_panelset()
```

.panelset[
.panel[.panel-name[app.R]

```r
hist(runif(100))
```
]

.panel[.panel-name[About]

Take a look at the R code in that other panel.
]
]
}\if{html}{\out{</div>}}
}

\section{Customizing styles}{
 The panelset uses custom CSS properties to make
it easier to change the styles of the panel tabs. The default values are
shown in the CSS code below. You can copy the whole CSS block to your
slides and modify the values to customize the style to fit your
presentation.\if{html}{\out{<div class="markdown">}}\preformatted{```\{css echo=FALSE\}
.panelset \{
  --panel-tabs-color: currentColor;
  --panel-tabs-color-active: currentColor;
  --panel-tabs-color-hover: currentColor;
  --panel-tabs-border-bottom: #ddd;
  --panel-tabs-inactive-opacity: 0.5;
  --panel-tabs-font-family: Menlo, Consolas, Monaco, Liberation Mono, Lucida Console, monospace;
\}
```
}\if{html}{\out{</div>}}
}

