% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide-tone.R
\name{slide_tone}
\alias{slide_tone}
\alias{use_slide_tone}
\alias{html_dependency_slide_tone}
\title{Slide Tone}
\usage{
use_slide_tone()

html_dependency_slide_tone()
}
\value{
An \code{htmltools::tagList()} with the slide tone dependencies, or an
\link[htmltools:htmlDependency]{htmltools::htmlDependency}.
}
\description{
Slide tone plays a subtle sound when you change slides.The tones increase in
pitch for each slide from a low C to a high C note. The tone pitch stays the
same for incremental slides.
}
\section{Functions}{
\itemize{
\item \code{use_slide_tone}: Adds slide tone to your xaringan slides.

\item \code{html_dependency_slide_tone}: Returns an \link[htmltools:htmlDependency]{htmltools::htmlDependency} with the tile
view dependencies. Most users will want to use \code{use_slide_tone()}.
}}

\section{Usage}{
 To add slide tone to your xaringan presentation, add the
following code chunk to your slides' R Markdown file.

\if{html}{\out{<div class="sourceCode markdown">}}\preformatted{```\{r xaringan-slide-tone, echo=FALSE\}
xaringanExtra::use_slide_tone()
```
}\if{html}{\out{</div>}}
}

\examples{
use_slide_tone()

}
\references{
\href{https://tonejs.github.io/}{tone.js}
}
