% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile-view.R
\name{tile_view}
\alias{tile_view}
\alias{use_tile_view}
\alias{html_dependency_tile_view}
\title{Tile View}
\usage{
use_tile_view()

html_dependency_tile_view()
}
\value{
An \code{htmltools::tagList()} with the tile view dependencies, or an
\code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}.
}
\description{
Tile view gives you a way to quickly jump between slides. Just press \strong{O} at
any point in your slideshow and the tile view appears. Click on a slide to
jump to the slide, or press \strong{O} to exit tile view.
}
\section{Functions}{
\itemize{
\item \code{use_tile_view()}: Adds tile view to your xaringan slides.

\item \code{html_dependency_tile_view()}: Returns an \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}} with the tile
view dependencies. Most users will want to use \code{use_tile_view()}.

}}
\section{Usage}{
 To add tile view to your xaringan presentation, add the
following code chunk to your slides' R Markdown file.

\if{html}{\out{<div class="sourceCode markdown">}}\preformatted{```\{r xaringan-tile-view, echo=FALSE\}
xaringanExtra::use_tile_view()
```
}\if{html}{\out{</div>}}
}

\examples{
use_tile_view()

}
