% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xaringanExtra-package.R
\docType{package}
\name{xaringanExtra-package}
\alias{xaringanExtra}
\alias{xaringanExtra-package}
\title{xaringanExtra: Extensions for xaringan}
\description{
\code{xaringanExtra} is a playground of enhancements and extensions for
\pkg{xaringan} slides.
}
\details{
\itemize{
\item Add an overview of your presentation with \code{\link[=use_tile_view]{use_tile_view()}}
\item Make your slide content editable with \code{\link[=use_editable]{use_editable()}}
\item Announce slide changes with a subtle tone: \code{\link[=use_slide_tone]{use_slide_tone()}}
\item Animate slide transitions with \code{\link[=use_animate_css]{use_animate_css()}}
\item Add tabbed panels to slides with \code{\link[=use_panelset]{use_panelset()}}
\item Add a logo to all of your slides with \code{\link[=use_logo]{use_logo()}}
\item Use the Tachyons CSS utility toolkit: \code{\link[=use_tachyons]{use_tachyons()}}
\item Add a live video feed to you slides with \code{\link[=use_webcam]{use_webcam()}}
\item Fit your slides to fill the browser window: \code{\link[=use_fit_screen]{use_fit_screen()}}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gadenbuie/xaringanExtra}
  \item Report bugs at \url{https://github.com/gadenbuie/xaringanExtra/issues}
}

}
\author{
\strong{Maintainer}: Garrick Aden-Buie \email{garrick@aden-buie.com} (\href{https://orcid.org/0000-0002-7111-0077}{ORCID})

Other contributors:
\itemize{
  \item Yotam Mann (tone.js) [contributor]
  \item Daniel Eden (animate.css) [contributor]
  \item  Tachyons authors [contributor]
  \item  Klaus Hartl, Fagner Brack, GitHub Contributors (js-cookie) [contributor]
  \item  Chris Andrejewski (himalaya) [contributor]
  \item Eric Londaits (text-poster.js) [contributor]
}

}
\keyword{internal}
