% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xaringanExtra-package.R
\docType{package}
\name{xaringanExtra-package}
\alias{xaringanExtra}
\alias{xaringanExtra-package}
\title{xaringanExtra: Extensions for xaringan}
\description{
\code{xaringanExtra} is a playground of enhancements and extensions for
\pkg{xaringan} slides.
}
\details{
\itemize{
\item Add an overview of your presentation with \code{\link[=use_tile_view]{use_tile_view()}}
\item Make your slide content editable with \code{\link[=use_editable]{use_editable()}}
\item Share your slides in style with \code{\link[=use_share_again]{use_share_again()}}
\item Broadcast your slides in real time to viewers with \code{\link[=use_broadcast]{use_broadcast()}}
\item Scribble on your slides during your presentation with \code{\link[=use_scribble]{use_scribble()}}
\item Announce slide changes with a subtle tone: \code{\link[=use_slide_tone]{use_slide_tone()}}
\item Animate slide transitions with \code{\link[=use_animate_css]{use_animate_css()}}
\item Add tabbed panels to slides with \code{\link[=use_panelset]{use_panelset()}}
\item Add a logo to all of your slides with \code{\link[=use_logo]{use_logo()}}
\item Add a top or bottom banner to all of your slides with \code{\link[=use_banner]{use_banner()}}
\item Add a search box to search through your slides with \code{\link[=use_search]{use_search()}}
\item Use the Tachyons CSS utility toolkit: \code{\link[=use_tachyons]{use_tachyons()}}
\item Add a live video feed to you slides with \code{\link[=use_webcam]{use_webcam()}}
\item Always play gifs from the start with \code{\link[=use_freezeframe]{use_freezeframe()}}
\item Add one-click code copying with \code{\link[=use_clipboard]{use_clipboard()}}
\item Fit your slides to fill the browser window: \code{\link[=use_fit_screen]{use_fit_screen()}}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pkg.garrickadenbuie.com/xaringanExtra/}
  \item \url{https://github.com/gadenbuie/xaringanExtra}
  \item Report bugs at \url{https://github.com/gadenbuie/xaringanExtra/issues}
}

}
\author{
\strong{Maintainer}: Garrick Aden-Buie \email{garrick@adenbuie.com} (\href{https://orcid.org/0000-0002-7111-0077}{ORCID})

Authors:
\itemize{
  \item Matthew T. Warkentin \email{warkentin@lunenfeld.ca} (\href{https://orcid.org/0000-0001-8730-3511}{ORCID}) (Contributed scribble)
}

Other contributors:
\itemize{
  \item Yotam Mann (tone.js) [copyright holder]
  \item Daniel Eden (animate.css) [copyright holder]
  \item  Tachyons authors [copyright holder]
  \item  Klaus Hartl, Fagner Brack, GitHub Contributors (js-cookie) [copyright holder]
  \item  Chris Andrejewski (himalaya) [copyright holder]
  \item Eric Londaits (text-poster.js) [copyright holder]
  \item Zeno Rocha (clipboard.js) [copyright holder]
  \item Nikita Karamov (shareon.js) [copyright holder]
  \item Ross Zurowski (fitvids.js) [copyright holder]
  \item  Michelle Bu and Eric Zhang (peerjs) [copyright holder]
  \item Kiril Vatev (tiny.toast) [copyright holder]
  \item André Restive (remark.search) [copyright holder]
  \item  Printio (Juriy Zaytsev, Maxim Chernyak) (fabric.js) [copyright holder]
  \item Christpher Antonellis (freezeframe.js) [copyright holder]
}

}
\keyword{internal}
